refactor: Treat DynamicLoader and GenieStructure as static definitions.
