+++ openage/convert/entity_object/conversion/CMakeLists.txt
@@ -4,8 +4,6 @@ add_py_modules(
 	combined_sound.py
 	combined_terrain.py
 	converter_object.py
-	dynamic_loader.py
-	genie_structure.py
 	modpack.py
 	stringresource.py
 )
