+++ openage/convert/value_object/read/CMakeLists.txt
@@ -1,5 +1,7 @@
 add_py_modules(
 	__init__.py
+	dynamic_loader.py
+	genie_structure.py
 	media_types.py
 	member_access.py
 	read_members.py
