+++ openage/convert/value_object/read/media/blendomatic.py
@@ -18,7 +18,7 @@ import numpy
 
 
 from .....log import dbg
-from ....entity_object.conversion.genie_structure import GenieStructure
+from ..genie_structure import GenieStructure
 
 
 if typing.TYPE_CHECKING:
