+++ openage/convert/value_object/read/media/datfile/civ.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 import typing
 
 from . import unit
-from .....entity_object.conversion.genie_structure import GenieStructure
+from ...genie_structure import GenieStructure
 from ....read.member_access import READ, READ_GEN, SKIP
 from ....read.read_members import MultisubtypeMember, EnumLookupMember
 from ....read.value_members import StorageType
