+++ openage/convert/value_object/read/media/datfile/sound.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 import typing
 
 
-from .....entity_object.conversion.genie_structure import GenieStructure
+from ...genie_structure import GenieStructure
 from ....read.member_access import READ_GEN, READ, SKIP
 from ....read.read_members import SubdataMember
 from ....read.value_members import StorageType
