+++ openage/convert/value_object/read/media/datfile/terrain.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 import typing
 
 
-from .....entity_object.conversion.genie_structure import GenieStructure
+from ...genie_structure import GenieStructure
 from ....read.member_access import READ, READ_GEN, SKIP
 from ....read.read_members import ArrayMember, SubdataMember, IncludeMembers
 from ....read.value_members import StorageType
